+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### 🔧 Backend / Logic
 - UAssetTool communication rewrite for a more stable and faster connection
-- Added detection for Hybrid IO Store Mods (Mods containing both raw and uasset files)
+- Added Hybrid mod install support (Merge raw assets and uasset files into a single IO Store bundle)
+- Added detection for Hybrid IO Store Mods
 
 ### 🎨 Frontend / UI
+- Added Hybrid bundle toggle to mod install panel, drop an hybrid-ready mod folder or merge paks and foldes to create one
 - Added a Sig Bypasser toggle in the Tools panel
 - Mod type labels are now additive
 - Dropping mods (pak/folders) into Mod Install panel will now add them to the installation queue
@@ -14,6 +16,6 @@
 - Added Hybrid label for Hybrid IO Store Mods
 - Added sorting for mods by modified date and name with asc/desc order
 - Added a resize handle for the filters and folders panel
-- Added support for Epic Games launcher and a toggle is added in settings to switch between steam and epic
+- Added support for Epic Games launcher and a toggle in settings to switch between steam and epic
 - Added Cyclops hero icon
 - Minor UI changes